changed foreign key type

+++ migrations/sqls/20220324021333-orders-table-up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE orders (
     id SERIAL PRIMARY KEY NOT NULL,
-    user_id bigint REFERENCES users(id),
+    user_id INTEGER REFERENCES users(id),
     status_of_order VARCHAR(255) NOT NULL
 );
